+++ classes/datamodel/player/config.lua
@@ -4,12 +4,12 @@ local playerConfig = {
 	-- skip the last to frames since idle has only 4 / 6
 	idle_frame_trim = 2,
 	sprite_offset_y = 1,
-	move_speed = 300,
+	move_speed = 150,
 	gamepad_deadzone = 0.2,
-	crosshair_max_distance = 60,
-	crosshair_radius = 6,
-	crosshair_line = 10,
-	hand_distance = 10,
+	crosshair_max_distance = 30,
+	crosshair_radius = 3,
+	crosshair_line = 5,
+	hand_distance = 5,
 	fire_gamepad_button = "rightshoulder"
 }