+++ main.lua
@@ -1,4 +1,5 @@
 Object = require "lib.classic"
+push = require "lib.push"
 
 local Game = require("classes.datamodel.game.init")
 local GameRenderer = require("classes.ui.gameRenderer")
@@ -6,6 +7,13 @@ local game
 local gameRenderer
 
 function love.load()
+	local screenW, screenH = love.window.getDesktopDimensions()
+	push:setupScreen(720, 448, screenW, screenH, {
+		fullscreen = true,
+		fullscreentype = "desktop",
+		resizable = false,
+	})
+
 	game = Game()
 	gameRenderer = GameRenderer(game)
 end
@@ -17,4 +25,8 @@ end
 
 function love.draw()
 	gameRenderer:draw()
+end
+
+function love.resize(w, h)
+	push:resize(w, h)
 end